Structure for Closure deps and Angular modules
