+++ build.json
@@ -3,10 +3,11 @@
   "src": [
     "node_modules/openlayers/src/**/*.js",
     "node_modules/openlayers/build/ol.ext/**/*.js",
-    "node_modules/ngeo/src/**/*.js"
+    "node_modules/ngeo/src/**/*.js",
+    "geoportailv3/static/js/**/*.js"
   ],
   "compile": {
-    "closure_entry_point": "main",
+    "closure_entry_point": "app_main",
     "externs": [
       "node_modules/openlayers/externs/bingmaps.js",
       "node_modules/openlayers/externs/closure-compiler.js",
@@ -75,6 +76,7 @@
     ],
     "compilation_level": "ADVANCED",
     "output_wrapper": "(function(){%output%})();",
-    "use_types_for_optimization": true
+    "use_types_for_optimization": true,
+    "manage_closure_dependencies": true
   }
 }
