+++ geoportailv3/static/js/appmodule.js
@@ -1,0 +1,8 @@
+/**
+ * @fileoverview This file defines the application's Angular module.
+ */
+goog.provide('app');
+
+(function() {
+  angular.module('app', ['ngeo', 'gettext']);
+})();
