+++ geoportailv3/templates/ngeo.html
@@ -28,8 +28,8 @@
 % endif
     <script>
       (function() {
-         var module = angular.module('app');
-         module.constant('langUrlTemplate', '/proj/build/locale/__lang__/geoportailv3.json');
+         var appModule = angular.module('app');
+         appModule.constant('langUrlTemplate', '/proj/build/locale/__lang__/geoportailv3.json');
        })();
     </script>
   </body>
